requirements.txt
discord.py
python-dotenv
pymongo
pytest
pytest-asyncio
pytest-sugar
mongomock